Define Resource and Status
Defines the fundamental types for the API. A `Resource` has a
unique identity, a `Status`, and the time of the last change. A
`Status` is nothing more than a choice between free, busy, and
occupied. The difference between busy and occupied is up to the
user, but generally can be thought of as busy can still take on
more, but occupied is at capacity.