Removes Sentinal Error in favor of private type
See Dave Cheney's excellent blog post on handling errors. Defines a
private error type along with a public function to check if the error is
an out of range error. It may be removed entirely in favor of just the
error message if no realistic use-cases for the sentinal error exist.