Adds initial store with Save method
Adds a simple boltDB store with a save method. Save method returns
errors when the store or database are uninitialized, when the Resource
does not have an ID, when the Resource is older than what is already in
the Store, and any time there are errors with the database or other
dependencies.