Updates resource tests for better time equality
Updates the `Resource` tests to better handle time equality. Since the
standard package `time.Time` values are not comparable with the `==`
operator, they must be compared with the `time.Time.Equal()` method
instead. This makes it such that the tests must explicitly check each
property of the `Resource` for equality. Since the `Resource` is a
fairly light structure, this is a minimal inconvenience for now, but may
become an issue in the future.