Handle distinct errors in the RESTful server
Adds an unexposed error type that can be used to provide additional
context for error conditions while serving http requests. In order to
simplify error handling for the RESTful server, http Handlers have been
updated to return an error. This error is used to return a correct http
error response based on the nature of the error.

Resolves #43