Return checkable errors from store
The store Save and Get methods have been updated to return errors that
can be checked to verify if they are bad data errors. The error checking
functions were updated to support the `Cause()` function from
github.com/pkg/errors so that the error when saving a resource could
correctly be checked.

Resolves #44