Check for Stale data errors from the store
The RESTful server now checks for stale data errors from the store.
While this works, I think there are some key refactorings that need to
take place before we can consider this complete.

The rest package should not be importing the store package in order to
check for a stale data error. The error check functions that are used
across packages may need to be moved out of the store package and into
the root package.

If the error check functions used for the store are going to be moved
into the root package, the Store interface may be best placed there.
Additionally, the private `errorCode` function used in the rest package
may be updated to utilize the exported `StaleError` and `ZeroValueError`
functions to map to correct status codes (409 and 400 respectively).