Not print functions in tests
There isn't really a great way to indicate which error check function was
run for the tests that were updated to use an error check function
instead of just a boolean. The error check functions are an important
test, a different strategy for checking that an error exists and is a
specific checkable error may be needed.